+++ tests/Model/Action/ActionTest.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace webignition\BasilParser\Tests\Model;
+namespace webignition\BasilParser\Tests\Model\Action;
 
-use webignition\BasilParser\Model\Action;
-use webignition\BasilParser\Model\ActionTypesInterface;
+use webignition\BasilParser\Model\Action\Action;
+use webignition\BasilParser\Model\Action\TypesInterface;
 
 class ActionTest extends \PHPUnit\Framework\TestCase
 {
     public function testCreate()
     {
-        $verb = ActionTypesInterface::RELOAD;
+        $verb = TypesInterface::RELOAD;
 
         $action = new Action($verb);
